payment details

+++ server/routes/main.js
@@ -9,6 +9,7 @@ const Campaign = require("../models/Campaign");
 const Creator = require("../models/Creator");
 const User = require("../models/User");
 const Wallet = require("../models/Wallet");
+const Payment = require("../models/Payment");
 
 const Bid = require("../models/Bid");
 const Withdraw = require("../models/Withdraw");
@@ -40,10 +41,10 @@ router.post("/add_order", async (req, res) => {
 
     const newOrder = new Order({
       ...req.body,
-      date: new Date(),
+      date: new Date()
     });
 
-    newOrder.save((err) => {
+    newOrder.save(err => {
       if (err) {
         console.log(err);
         res.status(400).json(`Error: ${err}`);
@@ -66,15 +67,15 @@ router.get("/get_orders", async (req, res) => {
       populate: [
         {
           path: "product",
-          model: "product",
-        },
-      ],
+          model: "product"
+        }
+      ]
     })
-    .then((p) => {
+    .then(p => {
       console.log(p);
       res.status(200).json(p);
     })
-    .catch((error) => {
+    .catch(error => {
       console.log(error);
       res.status(400).json(error);
     });
@@ -91,15 +92,15 @@ router.get("/get_orders/branduser/:userId", async (req, res) => {
       populate: [
         {
           path: "product",
-          model: "product",
-        },
-      ],
+          model: "product"
+        }
+      ]
     })
-    .then((p) => {
+    .then(p => {
       console.log(p);
       res.status(200).json(p);
     })
-    .catch((error) => {
+    .catch(error => {
       console.log(error);
       res.status(400).json(error);
     });
@@ -116,15 +117,15 @@ router.get("/get_orders/creatoruser/:userId", async (req, res) => {
       populate: [
         {
           path: "product",
-          model: "product",
-        },
-      ],
+          model: "product"
+        }
+      ]
     })
-    .then((p) => {
+    .then(p => {
       console.log(p);
       res.status(200).json(p);
     })
-    .catch((error) => {
+    .catch(error => {
       console.log(error);
       res.status(400).json(error);
     });
@@ -138,7 +139,7 @@ router.get(
     Order.find(
       {
         branduser: req.params.userId,
-        status: parseInt(req.params.status),
+        status: parseInt(req.params.status)
       },
       (err, orders) => {
         if (err) res.status(400).json(`Error: ${err}`);
@@ -156,7 +157,7 @@ router.get(
     Order.find(
       {
         creatoruser: req.params.userId,
-        status: parseInt(req.params.status),
+        status: parseInt(req.params.status)
       },
       (err, orders) => {
         if (err) res.status(400).json(`Error: ${err}`);
@@ -184,7 +185,7 @@ router.get(
 router.delete("/delete_order/:orderId", async (req, res) => {
   res.setHeader("Content-Type", "application/json");
 
-  Order.deleteOne({ _id: req.params.orderId }, (err) => {
+  Order.deleteOne({ _id: req.params.orderId }, err => {
     if (err) res.status(400).json(`Error: ${err}`);
     else res.status(200).send("Deleted one order successfully!");
   });
@@ -218,9 +219,9 @@ router.patch("/edit_order/:orderId", async (req, res) => {
     Order.updateOne(
       { _id: req.params.orderId },
       {
-        $set: req.body,
+        $set: req.body
       },
-      (err) => {
+      err => {
         if (err) res.status(400).json(`Error: ${err}`);
         else res.status(200).send("Edited an order");
       }
@@ -239,10 +240,10 @@ router.post("/add_product", (req, res) => {
   res.setHeader("Content-Type", "application/json");
 
   const newProduct = new Product({
-    ...req.body,
+    ...req.body
   });
 
-  newProduct.save((err) => {
+  newProduct.save(err => {
     console.log(err);
     if (err) res.status(400).json(`Error: ${err}`);
     else res.status(200).send("created a new product");
@@ -298,7 +299,7 @@ router.delete("/delete_product/:productId/:userId", async (req, res) => {
 
   Product.deleteOne(
     { _id: req.params.productId, userId: req.params.userId },
-    (err) => {
+    err => {
       if (err) res.status(400).json(`Error: ${err}`);
       else res.status(200).send("Deleted one product successfully!");
     }
@@ -326,10 +327,10 @@ router.post("/add_campaign", (req, res) => {
   res.setHeader("Content-Type", "application/json");
 
   const newCampaign = new Campaign({
-    ...req.body,
+    ...req.body
   });
 
-  newCampaign.save((err) => {
+  newCampaign.save(err => {
     if (err) {
       res.status(400).json(`Error: ${err}`);
     } else res.status(200).send("created a new campaign");
@@ -340,8 +341,8 @@ router.get("/get_all_campaigns", async (req, res) => {
   res.setHeader("Content-Type", "application/json");
   Campaign.find()
     .populate("product")
-    .then((p) => res.status(200).json(p))
-    .catch((error) => res.status(400).json(error));
+    .then(p => res.status(200).json(p))
+    .catch(error => res.status(400).json(error));
 });
 // Get all the campaigns of a user
 router.get("/get_campaigns/:userId", async (req, res) => {
@@ -359,7 +360,7 @@ router.delete("/delete_campaign/:campaignId/:userId", async (req, res) => {
 
   Campaign.deleteOne(
     { _id: req.params.campaignId, userId: req.params.userId },
-    (err) => {
+    err => {
       if (err) res.status(400).json(`Error: ${err}`);
       else res.status(200).send("Deleted one campaign successfully!");
     }
@@ -371,8 +372,8 @@ router.get("/get_campaign_by_id/:campaignId/:userId", async (req, res) => {
   res.setHeader("Content-Type", "application/json");
   Campaign.findOne({ _id: req.params.campaignId, userId: req.params.userId })
     .populate("product")
-    .then((order) => res.status(200).send(order))
-    .catch((error) => res.status(400).json(`Error: ${err}`));
+    .then(order => res.status(200).send(order))
+    .catch(error => res.status(400).json(`Error: ${err}`));
 });
 
 //Patch a campaign
@@ -382,9 +383,9 @@ router.patch("/edit_campaign/:orderId/:userId", async (req, res) => {
   Campaign.updateOne(
     { _id: req.params.orderId, userId: req.params.userId },
     {
-      $set: req.body,
+      $set: req.body
     },
-    (err) => {
+    err => {
       if (err) {
         console.log(err);
         res.status(400).json(`Error: ${err}`);
@@ -403,21 +404,87 @@ router.post("/add_creator", (req, res) => {
   res.setHeader("Content-Type", "application/json");
 
   const newCreator = new Creator({
-    ...req.body,
+    ...req.body
   });
 
-  newCreator.save((err) => {
+  newCreator.save(err => {
     console.log(err);
     if (err) res.status(400).json(`Error: ${err}`);
     else res.status(200).send("added a new creator");
   });
 });
 
+router.post("/add_payment", (req, res) => {
+  console.log(req.body);
+
+  res.setHeader("Content-Type", "application/json");
+
+  const newPayment = new Payment({
+    ...req.body
+  });
+
+  newPayment.save(err => {
+    if (err) res.status(400).json(`Error: ${err}`);
+    else res.status(200).send("added a new payment");
+  });
+});
+
+// Get all the campaigns of a user
+router.get("/get_payment_details/:email", async (req, res) => {
+  res.setHeader("Content-Type", "application/json");
+
+  Payment.find({ email: req.params.email }, (err, payment_info) => {
+    if (err) res.status(400).json(`Error: ${err}`);
+    else res.status(200).json(payment_info);
+  });
+});
+
+//Update by ID Method
+router.patch("/update_payment/:email", async (req, res) => {
+  try {
+    // const {name,street,city,zip,country} = req.body;
+
+    const result = await Payment.findOneAndUpdate(
+      { email: req.body.email },
+      {
+        $set: {
+          name: req.body.name,
+          street: req.body.street,
+          city: req.body.city,
+          zip: req.body.zip,
+          country: req.body.country
+        }
+      },
+      { new: true },
+      (err, doc) => {
+        if (err) {
+          console.log("Something wrong when updating data!");
+        }
+        res.send(doc);
+      }
+    );
+
+    // const result = await Payment.findByEmailAndUpdate(
+    //    {
+    //      "name":name,
+    //      "street":street,
+    //      "city":city,
+    //      "zip":zip,
+    //      "country":country
+    //    }
+    // )
+
+    // res.send(result)
+  } catch (error) {
+    res.status(400).json({ message: error.message });
+  }
+});
+
 router.get("/get_all_creators", async (req, res) => {
   res.setHeader("Content-Type", "application/json");
   Creator.find()
-    .then((p) => res.status(200).json(p))
-    .catch((error) => res.status(400).json(error));
+    .then(p => res.status(200).json(p))
+    .catch(error => res.status(400).json(error));
 });
 
 // Delete a campaign
@@ -426,7 +493,7 @@ router.delete("/delete_creator/:campaignId/:userId", async (req, res) => {
 
   Creator.deleteOne(
     { _id: req.params.campaignId, userId: req.params.userId },
-    (err) => {
+    err => {
       if (err) res.status(400).json(`Error: ${err}`);
       else res.status(200).send("Deleted one creator successfully!");
     }
@@ -438,11 +505,11 @@ router.get("/get_creator_by_id/:userId", async (req, res) => {
   console.log(req.params.userId);
   res.setHeader("Content-Type", "application/json");
   Creator.findOne({ userId: req.params.userId })
-    .then((order) => {
+    .then(order => {
       console.log(order);
       res.status(200).json(order);
     })
-    .catch((error) => {
+    .catch(error => {
       console.log(error);
       res.status(400).json(`Error: ${error}`);
     });
@@ -455,9 +522,9 @@ router.patch("/edit_creator/:orderId/:userId", async (req, res) => {
   Creator.updateOne(
     { _id: req.params.orderId, userId: req.params.userId },
     {
-      $set: req.body,
+      $set: req.body
     },
-    (err) => {
+    err => {
       if (err) res.status(400).json(`Error: ${err}`);
       else res.status(200).send("Edited a campaign");
     }
@@ -487,8 +554,8 @@ router.post("/add_user", (req, res) => {
         phoneNumber,
         funds: {
           amount: 0,
-          currency: "USD",
-        },
+          currency: "USD"
+        }
       });
       newUser.save((err, results) => {
         console.log(err);
@@ -502,22 +569,21 @@ router.post("/add_user", (req, res) => {
   });
 });
 
-
 // Users
 // Add a new user
 router.post("/add_wallet", (req, res) => {
-      const newWallet = new Wallet({
-       email: req.body.email,
-        funds: {
-          amount: req.body.funds.amount,
-          currency: "USD",
-        },
-      });
-      newWallet.save((err, results) => {
-        console.log(results)
-        if (err) res.status(400).json(`Error: ${err}`);
-        else res.status(200).send(results);
-      });
+  const newWallet = new Wallet({
+    email: req.body.email,
+    funds: {
+      amount: req.body.funds.amount,
+      currency: "USD"
+    }
+  });
+  newWallet.save((err, results) => {
+    console.log(results);
+    if (err) res.status(400).json(`Error: ${err}`);
+    else res.status(200).send(results);
+  });
 });
 
 // Get a single order
@@ -530,30 +596,28 @@ router.get("/get_wallet/:email", async (req, res) => {
   });
 });
 
-
 //Get All Users
 router.get("/get_all_users", (req, res) => {
   res.setHeader("Content-Type", "application/json");
   User.find({})
-    .then((users) => res.status(200).send(users))
-    .catch((error) => res.status(400).json(`Error: ${error}`));
+    .then(users => res.status(200).send(users))
+    .catch(error => res.status(400).json(`Error: ${error}`));
 });
 
 // Get a single user
 router.get("/get_user_by_id/:userId", async (req, res) => {
   res.setHeader("Content-Type", "application/json");
   User.findOne({ userId: req.params.userId })
-    .then((user) => res.status(200).send(user))
-    .catch((error) => res.status(400).json(`Error: ${err}`));
+    .then(user => res.status(200).send(user))
+    .catch(error => res.status(400).json(`Error: ${err}`));
 });
-
 
 // Get a single user
 router.get("/get_protifolio_by_id/:userId", async (req, res) => {
   res.setHeader("Content-Type", "application/json");
   Creator.findOne({ userId: req.params.userId })
-    .then((user) => res.status(200).send(user))
-    .catch((error) => res.status(400).json(`Error: ${err}`));
+    .then(user => res.status(200).send(user))
+    .catch(error => res.status(400).json(`Error: ${err}`));
 });
 
 //Patch a user
@@ -563,9 +627,9 @@ router.patch("/edit_user/:userId", async (req, res) => {
   User.updateOne(
     { userId: req.params.userId },
     {
-      $set: req.body,
+      $set: req.body
     },
-    (err) => {
+    err => {
       if (err) res.status(400).json(`Error: ${err}`);
       else res.status(200).send("Edited a user");
     }
@@ -576,14 +640,10 @@ router.patch("/edit_user/:userId", async (req, res) => {
 router.delete("/delete_user/:userId", async (req, res) => {
   res.setHeader("Content-Type", "application/json");
 
-  User.deleteOne(
-    { userId: req.params.userId },
-
-    (err) => {
-      if (err) res.status(400).json(`Error: ${err}`);
-      else res.status(200).send("Edited a user");
-    }
-  );
+  User.deleteOne({ userId: req.params.userId }, err => {
+    if (err) res.status(400).json(`Error: ${err}`);
+    else res.status(200).send("Edited a user");
+  });
 });
 
 // Add a new bid
@@ -604,10 +664,10 @@ router.post("/add_bid", (req, res) => {
           {
             creator,
             creatorLevel,
-            price,
-          },
+            price
+          }
         ],
-        winningBid: {},
+        winningBid: {}
       });
       newBid.save((err, results) => {
         console.log(err);
@@ -616,13 +676,13 @@ router.post("/add_bid", (req, res) => {
         else res.status(200).send(results);
       });
     } else {
-      let index = bid.creators.findIndex((c) => c.creator == creator);
+      let index = bid.creators.findIndex(c => c.creator == creator);
 
       if (index !== -1) {
         bid.creators[index] = {
           creator,
           creatorLevel,
-          price,
+          price
         };
       } else {
         bid.creators.push({ creator, creatorLevel, price });
@@ -644,8 +704,8 @@ module.exports = router;
 router.get("/get_all_bids", async (req, res) => {
   res.setHeader("Content-Type", "application/json");
   Bid.find()
-    .then((p) => res.status(200).json(p))
-    .catch((error) => res.status(400).json(error));
+    .then(p => res.status(200).json(p))
+    .catch(error => res.status(400).json(error));
 });
 
 router.get("/get_bid_by_campaign/:campaignId", async (req, res) => {
@@ -653,11 +713,11 @@ router.get("/get_bid_by_campaign/:campaignId", async (req, res) => {
   res.setHeader("Content-Type", "application/json");
   Bid.findOne({ campaign: req.params.campaignId })
     .populate("campaign creators.creator")
-    .then((p) => {
+    .then(p => {
       console.log(p);
       res.status(200).json(p);
     })
-    .catch((error) => {
+    .catch(error => {
       console.log(error);
       res.status(400).json(error);
     });
@@ -666,8 +726,8 @@ router.get("/get_bid_by_campaign/:campaignId", async (req, res) => {
 router.get("/get_bids_by_creator/:creatorId", async (req, res) => {
   res.setHeader("Content-Type", "application/json");
   Bid.find({ creators: { $elemMatch: { creator: req.params.creatorId } } })
-    .then((p) => res.status(200).json(p))
-    .catch((error) => res.status(400).json(error));
+    .then(p => res.status(200).json(p))
+    .catch(error => res.status(400).json(error));
 });
 
 router.patch("/edit_bid/:bidId", async (req, res) => {
@@ -677,10 +737,10 @@ router.patch("/edit_bid/:bidId", async (req, res) => {
     { _id: req.params.bidId },
     {
       $set: {
-        winningBid: req.body.winningBid,
-      },
+        winningBid: req.body.winningBid
+      }
     },
-    (err) => {
+    err => {
       if (err) res.status(400).json(`Error: ${err}`);
       else res.status(200).send("Edited a bid");
     }
@@ -693,7 +753,7 @@ router.delete("/delete_bid/:bidId", async (req, res) => {
 
   Bid.deleteOne(
     { _id: req.params.campaignId, userId: req.params.userId },
-    (err) => {
+    err => {
       if (err) res.status(400).json(`Error: ${err}`);
       else res.status(200).send("Deleted one creator successfully!");
     }
